Modify BSLWexport.pl CLI options.
Instead of assuming $ARGV[0] is the configuration file, we now take
an explicit -c to indicate that the next argument is the configuration
file.

We also now take a -f to indicate that the next argument is an already
dumped file to upload.  This option is useful if you had a run that
failed to upload for some reason, but you got the MARC output.

The code of BSLWexport.pl is also modified to do something logical when
the new arguments are set.

Signed-off-by: Jason Stephenson <jstephenson@mvlc.org>