Make some fixups to Backstage::Import::date_comp.
Wrap DateTime::Format::ISO8601->parse_datetime() calls in eval blocks
to catch any parse errors.  If there are parse errors, carp about the
problem, make sure that date is undefined, and move on.  We've had
the import crash with some bad dates in our records.

Also, alter the return value logic a little bit in that function.
Assume that we don't need an import, then add an elsif for the
BSLW record date being defined and the own record not defined
to set the return value to 1, indicating an import is needed.

Finally remove an extraneous newline from a carp statement
higher up in the file.

Signed-off-by: Jason Stephenson <jstephenson@mvlc.org>