Changes so that emails can be sent without authentication.
If the user and/or password are not defined or empty in the email
smtp preferences, email will be sent without authenticating to the
mail server. Also, if the port setting isn't set , it will default
to 25, and the encryption setting will default to 'none' if it is
missing or not set.

These changes were made because there is not point authenticating
if your computer is allowed to relay through your mail server,
anyway.

Signed-off-by: Jason Stephenson <jstephenson@mvlc.org>